Removed all assets except for one .png image

+++ src/engine/core/App.h
@@ -21,7 +21,7 @@ struct GraphicsConfig {
     uint32_t logical_resolution_height = 648;
     uint32_t target_fps = 60;
     bool vsync_enabled = true;
-    float pixels_per_meter = 64.0f;
+    uint32_t pixels_per_meter = 64;
 };
